fix(GH-901): align citizenapi unit tests with removed getOffices enrichment
Drop unused /source stubs from appointment controller tests, pass resolveReferences on reserve POST mocks, enrich scope providers from the same source fetch, and fix UpdateService HTTP call expectations.

+++ zmscitizenapi/src/Zmscitizenapi/Services/Core/ZmsApiFacadeService.php
@@ -471,7 +471,7 @@ class ZmsApiFacadeService
             slotsPerAppointment: ((string) $matchingScope->getSlotsPerAppointment() === '' ? null : (string) $matchingScope->getSlotsPerAppointment()),
             appointmentsPerMail: ((string) $matchingScope->getAppointmentsPerMail() === '' ? null : (string) $matchingScope->getAppointmentsPerMail()),
             whitelistedMails: ((string) $matchingScope->getWhitelistedMails() === '' ? null : (string) $matchingScope->getWhitelistedMails()),
-            reservationDuration: (int) MapperService::extractReservationDuration($matchingScope),
+            reservationDuration: MapperService::extractReservationDuration($matchingScope),
             activationDuration: MapperService::extractActivationDuration($matchingScope),
             hint: ((string) $matchingScope->getScopeHint() === '' ? null : (string) $matchingScope->getScopeHint())
         );
